+++ .gitignore
@@ -6,5 +6,6 @@ __pycache__
 *idfsim
 *.swp
 trec_eval.9.0/
+trec_eval-9.0.5
 *.pt
 text/
