+++ datasets/trecqa.py
@@ -1,40 +1,22 @@
-import math
 import os
 
-import numpy as np
 import torch
 from torchtext.data.example import Example
 from torchtext.data.field import Field
 from torchtext.data.iterator import BucketIterator
-from torchtext.data.pipeline import Pipeline
 from torchtext.vocab import Vectors
 
 from datasets.castor_dataset import CastorPairDataset
 from datasets.idf_utils import get_pairwise_word_to_doc_freq, get_pairwise_overlap_features
 
 
-def get_class_probs(sim, *args):
-    """
-    Convert a single label into class probabilities.
-    """
-    class_probs = np.zeros(MSRVID.NUM_CLASSES)
-    ceil, floor = math.ceil(sim), math.floor(sim)
-    if ceil == floor:
-        class_probs[floor] = 1
-    else:
-        class_probs[floor] = ceil - sim
-        class_probs[ceil] = sim - floor
-
-    return class_probs
-
-
-class MSRVID(CastorPairDataset):
-    NAME = 'msrvid'
-    NUM_CLASSES = 6
-    ID_FIELD = Field(sequential=False, use_vocab=False, batch_first=True)
+class TRECQA(CastorPairDataset):
+    NAME = 'trecqa'
+    NUM_CLASSES = 2
+    ID_FIELD = Field(sequential=False, tensor_type=torch.FloatTensor, use_vocab=False, batch_first=True)
     TEXT_FIELD = Field(batch_first=True, tokenize=lambda x: x)  # tokenizer is identity since we already tokenized it to compute external features
     EXT_FEATS_FIELD = Field(tensor_type=torch.FloatTensor, use_vocab=False, batch_first=True, tokenize=lambda x: x)
-    LABEL_FIELD = Field(sequential=False, tensor_type=torch.FloatTensor, use_vocab=False, batch_first=True, postprocessing=Pipeline(get_class_probs))
+    LABEL_FIELD = Field(sequential=False, use_vocab=False, batch_first=True)
 
     @staticmethod
     def sort_key(ex):
@@ -42,20 +24,20 @@ class MSRVID(CastorPairDataset):
 
     def __init__(self, path):
         """
-        Create a MSRVID dataset instance
+        Create a TRECQA dataset instance
         """
-        super(MSRVID, self).__init__(path)
+        super(TRECQA, self).__init__(path)
 
     @classmethod
-    def splits(cls, path, train='train', test='test', **kwargs):
-        return super(MSRVID, cls).splits(path, train=train, test=test, **kwargs)
+    def splits(cls, path, train='train-all', validation='raw-dev', test='raw-test', **kwargs):
+        return super(TRECQA, cls).splits(path, train=train, validation=validation, test=test, **kwargs)
 
     @classmethod
     def iters(cls, path, vectors_name, vectors_cache, batch_size=64, shuffle=True, device=0, vectors=None, unk_init=torch.Tensor.zero_):
         """
         :param path: directory containing train, test, dev files
         :param vectors_name: name of word vectors file
-        :param vectors_cache: path to word vectors file
+        :param vectors_cache: directory containing word vectors file
         :param batch_size: batch size
         :param device: GPU device
         :param vectors: custom vectors - either predefined torchtext vectors or your own custom Vector classes
@@ -65,8 +47,8 @@ class MSRVID(CastorPairDataset):
         if vectors is None:
             vectors = Vectors(name=vectors_name, cache=vectors_cache, unk_init=unk_init)
 
-        train, test = cls.splits(path)
+        train, validation, test = cls.splits(path)
 
-        cls.TEXT_FIELD.build_vocab(train, test, vectors=vectors)
+        cls.TEXT_FIELD.build_vocab(train, validation, test, vectors=vectors)
 
-        return BucketIterator.splits((train, test), batch_size=batch_size, repeat=False, shuffle=shuffle, device=device)
+        return BucketIterator.splits((train, validation, test), batch_size=batch_size, repeat=False, shuffle=shuffle, device=device)
