+++ mp_cnn/dataset.py
@@ -1,30 +1,14 @@
-from collections import defaultdict
-from enum import Enum
-import math
 import os
 
-import numpy as np
 import torch
-from torch.autograd import Variable
 import torch.nn as nn
-import torch.utils.data as data
 
 from datasets.sick import SICK
 from datasets.msrvid import MSRVID
-
-# logging setup
-import logging
-logger = logging.getLogger(__name__)
-logger.setLevel(logging.INFO)
-
-ch = logging.StreamHandler()
-ch.setLevel(logging.DEBUG)
-formatter = logging.Formatter('%(levelname)s - %(message)s')
-ch.setFormatter(formatter)
-logger.addHandler(ch)
+from datasets.trecqa import TRECQA
 
 
-class UnknownWorcVecCache(object):
+class UnknownWordVecCache(object):
     """
     Caches the first randomly generated word vector for a certain size to make it is reused.
     """
@@ -47,7 +31,7 @@ class MPCNNDatasetFactory(object):
     def get_dataset(dataset_name, word_vectors_dir, word_vectors_file, batch_size, device):
         if dataset_name == 'sick':
             dataset_root = os.path.join(os.pardir, os.pardir, 'data', 'sick/')
-            train_loader, dev_loader, test_loader = SICK.iters(dataset_root, word_vectors_file, word_vectors_dir, batch_size, device=device, unk_init=UnknownWorcVecCache.unk)
+            train_loader, dev_loader, test_loader = SICK.iters(dataset_root, word_vectors_file, word_vectors_dir, batch_size, device=device, unk_init=UnknownWordVecCache.unk)
             embedding_dim = SICK.TEXT_FIELD.vocab.vectors.size()
             embedding = nn.Embedding(embedding_dim[0], embedding_dim[1])
             embedding.weight = nn.Parameter(SICK.TEXT_FIELD.vocab.vectors)
@@ -55,11 +39,20 @@ class MPCNNDatasetFactory(object):
         elif dataset_name == 'msrvid':
             dataset_root = os.path.join(os.pardir, os.pardir, 'data', 'msrvid/')
             dev_loader = None
-            train_loader, test_loader = MSRVID.iters(dataset_root, word_vectors_file, word_vectors_dir, batch_size, device=device, unk_init=UnknownWorcVecCache.unk)
+            train_loader, test_loader = MSRVID.iters(dataset_root, word_vectors_file, word_vectors_dir, batch_size, device=device, unk_init=UnknownWordVecCache.unk)
             embedding_dim = MSRVID.TEXT_FIELD.vocab.vectors.size()
             embedding = nn.Embedding(embedding_dim[0], embedding_dim[1])
             embedding.weight = nn.Parameter(MSRVID.TEXT_FIELD.vocab.vectors)
             return MSRVID, embedding, train_loader, test_loader, dev_loader
+        elif dataset_name == 'trecqa':
+            if not os.path.exists('../utils/trec_eval-9.0.5/trec_eval'):
+                raise FileNotFoundError('TrecQA requires the trec_eval tool to run. Please run get_trec_eval.sh inside Castor/utils (as working directory) before continuing.')
+            dataset_root = os.path.join(os.pardir, os.pardir, 'data', 'TrecQA/')
+            train_loader, dev_loader, test_loader = TRECQA.iters(dataset_root, word_vectors_file, word_vectors_dir, batch_size, device=device, unk_init=UnknownWordVecCache.unk)
+            embedding_dim = TRECQA.TEXT_FIELD.vocab.vectors.size()
+            embedding = nn.Embedding(embedding_dim[0], embedding_dim[1])
+            embedding.weight = nn.Parameter(TRECQA.TEXT_FIELD.vocab.vectors)
+            return TRECQA, embedding, train_loader, test_loader, dev_loader
         else:
             raise ValueError('{} is not a valid dataset.'.format(dataset_name))
